removed parent list mixin
