+++ tests/unit/mixins/all.ts
@@ -1,4 +1,3 @@
 import './createCssTransitionMixin';
 import './createFormFieldMixin';
-import './createParentListMixin';
 import './createVNodeEvented';
